refactor: Standardize response properties across models; remove WeatherForecastController; enhance BankRepository and UserService with new methods

+++ API_BANKING_PAYMENT/Respositories/IRepositories/IRepository.cs
@@ -6,7 +6,7 @@ namespace API_BANKING_PAYMENT.Respositories.IRepositories
      
         public interface IRepository<T> where T : class
         {
-         Task<T> GetById(int id);
+         Task<T> GetById(long id);
          Task<IEnumerable<T>> GetAll();
          Task<bool> Add(T entity);
          Task<bool> Update(T entity);
